added functionality up to user registration

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -11,6 +13,10 @@ from .forms import JobForm
 
 # LOGIN VIEW
 def loginPage(request):
+
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -29,7 +35,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Username OR Password does not exist')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 # LOGOUT VIEW
@@ -37,7 +43,12 @@ def logoutUser(request):
     logout(request)
     return redirect('home')
 
+# REGISTER PAGE VIEW
+def registerPage(request):
+    page = 'regiser'
+    return render(request, 'base/login_register.html')
 
+# HOME VIEW
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     
@@ -52,11 +63,13 @@ def home(request):
     context = {'jobs': jobs, 'companies': companies, 'job_count': job_count}
     return render(request, 'base/home.html', context)
 
+# JOB VIEW
 def job(request, pk):
     job = Job.objects.get(id=pk)
     context = {'job': job}        
     return render(request, 'base/job.html', context)
 
+@login_required(login_url='login')
 def createJob(request):
     form = JobForm()
     if request.method == 'POST':
@@ -67,9 +80,14 @@ def createJob(request):
     context = {'form': form}
     return render(request, 'base/job_form.html', context)
 
+@login_required(login_url='login')
 def updateJob(request, pk):
     job = Job.objects.get(id=pk)
     form = JobForm(instance=job)
+
+    # restrict others from updating others Jobs
+    if request.user != Job.host:
+        return HttpResponse('You are not allowed here')
 
     if request.method == 'POST':
         form = JobForm(request.POST, instance=job)
@@ -80,8 +98,14 @@ def updateJob(request, pk):
     context = {'form': form}
     return render(request, 'base/job_form.html', context)
 
+@login_required(login_url='login')
 def deleteJob(request, pk):
     job = Job.objects.get(id=pk)
+
+    # restrict others from deleting others Jobs
+    if request.user != Job.host:
+        return HttpResponse('You are not allowed here')
+    
     if request.method == 'POST':
         job.delete()
         return redirect('home')
